requirements.txt
asgiref==3.8.1
autocommand==2.2.2
backports.tarfile==1.2.0
Django==4.2.16
django-ckeditor==6.7.1
django-js-asset==2.2.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
more-itertools==10.5.0
pillow==10.4.0
platformdirs==4.3.6
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.1
typeguard==4.3.0
typing_extensions==4.12.2
zipp==3.20.2
